[FIX] website: blog tags do not get removed on refresh or search
Before the change in the blog page of a website every tag but one used to disappear when refreshing the page or when the search bar was used.
I changed the blog's tags template to implement a prevent_redirect parameter similar to the one used in PR #164577, fixing the tag behaviour.
The implementation of a prevent_redirect parameter is necessary to keep tag selection as a POST while allowing GET requests from crawlers (again, see PR #164577).

Fixes #164577