[IMP] mail: one partner find or create to rule them all !
In order to unify methods searching or creating partners, let us create
yet another one. Introducing '_partner_find_from_emails' that should
act as the main find, sort and create partner record-based tool method.

It relies on Partner._find_or_create_from_emails() for finding and
creating, as it is already covering a lot of use cases (parsing name
and email, multi-email input, ...). It is improved to support a sorting
key so that it is now possible to choose which partner matches a given
input when multiple results match.

Next commits are going to update codebase to use that tool.

Task-
Prepares Task-