[REF] web: ControlPanel buttons as dropdown on small screens
On smaller screens, the ControlPanel's main buttons are grouped as a
dropdown if there are more than one (to take less space).

This commit refactors the logic behind this behaviour to have less
juggling with the DOM and the buttons' classes and use the
`disableDropdown` state to hide them when needed; also it uses the
`.o-dropdown-item-unstyled-button` to cleanup the buttons that are
transformed into dropdown items.