[IMP] mail, sms: log pre-sending errors in the chatter
When sending messages in mass (by email or sms), the messages were logged in
the chatter no matter if they were sent or not. The user has no mean to know if
the system has even try to send the mail or not as it appears as sent in the
chatter no matter the circumstance (because no notifications are created when
sending in mass).

When we know before sending the message that there is an error, instead of
logging the message in the chatter, we now log an error (ex.: blacklisted mail,
malformed phone number, ...). With this change, the user will at least know
that some message weren't sent to the user because there was something wrong in
the records. Unfortunately, if an error happens later, the message will still
be logged in the chatter with no mean for the user to know if the message was
sent or not.

Task-2834862