[IMP] sale: try to set `reversed_entry_id` when creating credit note
In the context of a sale order with products invoiced on delivery,
if there's a refund, it is possible to create a credit note from the
sale order (by clicking "Create Invoice" button).

From a user perspective, it would be logic if you created an invoice, to
have the credit note being the reversal of this invoice.
Here the goal is to set when a `reversed_entry_id` for this simple case,
and nothing otherwise.

task-3862316