[FIX] account: autoinstall first matching coa
When new modules with coas are being installed, the matching
coa based on country is autoinstalled if possible, but if
multiple charts match then we overwrite the autoinstall
and use the one from the last module.
Here we make sure we do not overwrite the chart if it is not
associated with the company's country.