[IMP] web: display current time in datetime field
Previously when clicking on a date in datetime picker we
are getting only the current hour but not the current minute.
User should get the current time while clicking on the date
so that he can directly apply that time.

After this commit the current datetime will be directly
be applied.