[FIX] rating : remove one click rating behavior
Context:
In its current form, the HTTP route using `/rate/<string:token>/<int:rate>`,
that is used in URL links generated by emails asking for customer feedback,
will have a "one-click write" behavior that commits the rating encoded
in `<int:rate>`. So the record is already updated with the new value.
Only then renders the feedback submit Form, where the user
can change their preselected rating and leave a commentary before submitting
it via a POST request.

Over the last years email providers and third party services will
preemptively crawl URL in emails to verify if they are "dangerous" (e.g.
Microsoft's Safe link and Defender fatures).
By doing so, the rating for the records will be set to the last clicked
URL using the above HTTP route, without any input from a human user.

It will be creating a number of false positive or negative ratings that do not
represent the real satisfaction of the surveyed customer.

Proposed solution:
We are removing the implicit "one-click write" behavior of the route and
favoring the Form submission expecting human input.
This will increase the likelihood a given rating was the result of the customer
clicking on the link and taking the time to leave a feedback, versus
random inputs from URL bots.

Future outlook:
With the ever-changing landscape of URL crawler bots, trying to develop
counter measures to identify and flag bot behavior is possible, but would
add immense overhead and maintenance cost of the code for an arguably minor
feature, i.e. the customer can rate the ticket by just clicking the picture
in the email.
Other methods to make these routes more resilient to URL crawlers might
be revisited in the future, but for the short-term this seems like a sufficient
solution to the unwanted behavior.

OPW-3411799
OPW-4292166