[FIX] account: filter journal dashboard
On the journal dashboard, when clinking on the name
of the sale/purchase journals, we filter the credit notes/refunds and
we show only invoices/bills. But when clicking on the informations links ('To
validate', 'To pay', etc) the filter shouldn't be active since we
display the number regarding both invoice and refund.
This is due to the fact that all links are linked to the same action,
which has the default filter in its context.

With this commit we add a context key on the informations links to
disable the filter.

Steps:

- Have 1 draft bill and 1 draft refund
- Go to journal dashboard
- On the Vendor Bills journal, click on '2 To Validate'
-> only the bill is shown, same goes for 'To pay', 'Late' and for the
   Customer Invoices journal.

opw-4328137