[IMP] sales_team: activate multi from form
Purpose
=======
Activate the multi team option in one click from the crm team
and crm team member form views.

Technical:
As the alert is displayed when the form isn't saved (computed text field),
adding the option link directly in the xml with a form controller to catch
the click and activate the option.

The "Configuration" menu in crm is only available if the user has the
"sales_team.group_sale_manager" group, the Multi-Team option in the settings
is also only accessible when having this group.
=> No need to add a group check on the button as people having access to the
form views also have the multi-team option right.
Just checking the group in the js form controller to make sure we only really try
to activate if the user has the right.

Adding an "invisible" attribute on the crm team form header to make sure the header
is present in the hierachy only if the action button is visible. Having an empty header
was adding a strange space on top of the form sheet which was misaligning it with the chatter.

Task-4283552