[FIX] web: avoid processing the embedded actions when no embedded action
Before this commit, the condition to search in the embedded actions of
the action to call the one to load instead of the main one (if the
current user re-ordered the actions displayed in the top bar) just check
if `embedded_action_ids` key exists in the action loaded but it does not
check the value of that attribute.

This commit checks if the list inside that attribute contains at least
one embedded action before searching the embedded action to load.

task-4295738