[PERF] website_sale_collect: stop creating useless sales orders
When click and collect is enabled, users can see pickup locations from
a location selector on /product page. Before this commit a sales order
was created on opening location selector that could result in a lot of
abandoned carts. After this commit temporary orders are created instead
on opening location selector and a cart is created only after selecting
a pickup location.