[FIX] website_blog: allow editing of the blog page title
Following [1], the blog page reverted to using the first blog post as
the cover. However, the title of the page ("Our Latest Posts") was made
non-editable and, consequently, non-translatable, which is not ideal.

This commit addresses the issue by making the title editable, ensuring
it can also be translated.

[1]: https://github.com/odoo/odoo/commit/05ef95d3f13ac42713bb8d8a3002f149345cc08b

opw-4289735