[FIX] website_sale: cart entries will not overflow in mobile eCommerce page
Before the change, in the mobile version of the eCommerce cart, if a product had a name that was too long or the font size was too big it would push the elements to its right offscreen.
After the change, the overflow of a product name that is too long or too big gets replaced with ellipsis, keeping the elements to its right from beign pushed offscreen.

opw-4253880