[FIX] html_editor: don't use self closing `a` elements in editor
Issue:
======
Self-closing `a` elements doesn't render properly.

Steps to reproduce the issue:
=============================
- Go to email templates
- Open "Send Quotation" template
- You will see the `a` element of the attachment is appearing a lot of
  times.

Origin of the issue:
====================
Since the content of the template is in xml, using self closing `a` element
is acceptable but it's not valid html so using the same content of xml
into html will produce incorrect ui.

Solution:
=========
For now we take the self closing `a` elements and add an empty text
node inside them.

opw-4322575