[IMP] core: ensure constraint for required boolean fields
When we add a required boolean field with default `False` the ORM
deoesn't fill existing rows and thus outputs an error to the logs.

Example model to reproduce the issue in a custom Module `mymodule`:
```
class A(models.Model):
    _inherit = "res.users"
    foo = field.Boolean(required=True, default=False)
```

If we install `mymodule`
```
2024-11-19 13:05:31,076 277801 INFO test_m_x odoo.registry: module mymodule: creating or updating database tables
2024-11-19 13:05:31,092 277801 ERROR test_m_x odoo.schema: Table 'res_users': unable to set NOT NULL on column 'foo'
```
The ORM fails to add the `NOT NULL` and moves on (no crash).

In this patch we ensure that the ORM adds the `NOT NULL` constraint.