[FIX] web_editor: replace $0 with direct inherited view content
The $0 placeholder in templates is highly useful, but it is not
considered when saving a view. For instance, the search icon in the shop
search bar has an incorrect xpath:
`/data/xpath[3]/form/t[1]/div[1]/button[1]/i[1]`, where the inner `t[1]`
actually points to another view through a `<t>$0</t>` containing the
inherited view.

The method responsible for handling this replacement fails when the
element is editable, causing a crash.

This commit replaces any $0 placeholders with their respective inherited
views, ensuring the xpath is correctly resolved and targeted.

task-3609835