[FIX] l10n_din5008: duplicate title in expense reports
Issue:
Expense title is duplicated when printing an expense report for localizations using the DIN5008 standard.

Steps to reproduce:
- Add the Austrian localization to the DB (or any other using DIN5008 - e.g. DE, CH)
- Add the Expenses module to the DB
- Switch to the Austrian company
- Create a new employee
- Create a new expense report
- Print the PDF of the report

Cause:
DIN5008 reports have a specific field in their header for the title of the document.
As such, the title present in the body of the page should not be displayed.
Because of oversight, this is not the case.

Solution:
Apply styling to "not display" the title present in the body of the report.

Ticket:
4314414